+++ Sources/Conduit/Auth/TokenStorage/OAuth2TokenMemoryStore.swift
@@ -12,6 +12,7 @@ import Foundation
 public class OAuth2TokenMemoryStore: OAuth2TokenStore {
 
     private var tokens: [String: OAuth2Token & DataConvertible] = [:]
+    private var refreshTokenLockExpirations: [String: Date] = [:]
 
     /// Creates a new OAuth2TokenMemoryStore
     public init() {}
@@ -33,6 +34,23 @@ public class OAuth2TokenMemoryStore: OAuth2TokenStore {
     public func tokenFor<Token: OAuth2Token & DataConvertible>(client: OAuth2ClientConfiguration,
                                                                authorization: OAuth2Authorization) -> Token? {
         return tokens[tokenKeyFor(client: client, authorization: authorization)] as? Token
+    }
+
+    public func lockRefreshToken(timeout: TimeInterval, client: OAuth2ClientConfiguration, authorization: OAuth2Authorization) -> Bool {
+        let tokenKey = tokenKeyFor(client: client, authorization: authorization)
+        refreshTokenLockExpirations[tokenKey] = Date(timeIntervalSinceNow: timeout)
+        return true
+    }
+
+    public func unlockRefreshTokenFor(client: OAuth2ClientConfiguration, authorization: OAuth2Authorization) -> Bool {
+        let tokenKey = tokenKeyFor(client: client, authorization: authorization)
+        refreshTokenLockExpirations[tokenKey] = nil
+        return true
+    }
+
+    public func refreshTokenLockExpirationFor(client: OAuth2ClientConfiguration, authorization: OAuth2Authorization) -> Date? {
+        let tokenKey = tokenKeyFor(client: client, authorization: authorization)
+        return refreshTokenLockExpirations[tokenKey]
     }
 
 }