+++ Tests/ConduitTests/Auth/OAuth2RequestPipelineMiddlewareTests.swift
@@ -9,6 +9,29 @@
 import XCTest
 @testable import Conduit
 
+class CustomRefreshTokenGrantStrategy: OAuth2TokenGrantStrategy {
+
+    enum Error: Swift.Error {
+        case didExecute
+    }
+
+    func issueToken(completion: @escaping (Result<BearerToken>) -> Void) {
+        completion(.error(CustomRefreshTokenGrantStrategy.Error.didExecute))
+    }
+
+    func issueToken() throws -> BearerToken {
+        return BearerToken(accessToken: "", expiration: Date())
+    }
+
+}
+
+struct CustomRefreshTokenGrantStrategyFactory: OAuth2RefreshStrategyFactory {
+    func make(refreshToken: String, clientConfiguration: OAuth2ClientConfiguration) -> OAuth2TokenGrantStrategy {
+        return CustomRefreshTokenGrantStrategy()
+    }
+}
+
+// swiftlint:disable type_body_length
 class OAuth2RequestPipelineMiddlewareTests: XCTestCase {
 
     let validClientID = "test_client"
@@ -91,9 +114,77 @@ class OAuth2RequestPipelineMiddlewareTests: XCTestCase {
                 }
                 let authorizationHeader = request.allHTTPHeaderFields?["Authorization"]
                 XCTAssertTrue(authorizationHeader?.contains("Bearer") == true)
-                XCTAssertNotEqual(authorizationHeader, expiredToken.accessToken)
+                XCTAssert(authorizationHeader?.contains(expiredToken.accessToken) == false)
 
                 refreshTokenExpectation.fulfill()
+            }
+        }
+
+        waitForExpectations(timeout: 2)
+    }
+
+    func testAllowsCustomTokenRefreshGrants() throws {
+        let authorization = OAuth2Authorization(type: .bearer, level: .user)
+        let validClientConfiguration = try makeValidClientConfiguration()
+        let request = try makeDummyRequest()
+        let tokenStorage = OAuth2TokenMemoryStore()
+        var sut = OAuth2RequestPipelineMiddleware(clientConfiguration: validClientConfiguration, authorization: authorization, tokenStorage: tokenStorage)
+
+        sut.refreshStrategyFactory = CustomRefreshTokenGrantStrategyFactory()
+
+        let refreshTokenExpectation = expectation(description: "token refreshed")
+
+        let clientCredentialsStrategy = OAuth2ClientCredentialsTokenGrantStrategy(clientConfiguration: validClientConfiguration)
+        clientCredentialsStrategy.issueToken { result in
+            guard let token = result.value else {
+                XCTFail("No token")
+                return
+            }
+            let expiredToken = BearerToken(accessToken: token.accessToken, refreshToken: token.refreshToken, expiration: Date())
+
+            tokenStorage.store(token: expiredToken, for: validClientConfiguration, with: authorization)
+
+            sut.prepareForTransport(request: request) { result in
+                guard result.error is CustomRefreshTokenGrantStrategy.Error else {
+                    XCTFail("Custom strategy ignored")
+                    return
+                }
+
+                refreshTokenExpectation.fulfill()
+            }
+        }
+
+        waitForExpectations(timeout: 2)
+    }
+
+    func testEmptyTokenRefreshStrategyPreventsRefreshes() throws {
+        let authorization = OAuth2Authorization(type: .bearer, level: .user)
+        let validClientConfiguration = try makeValidClientConfiguration()
+        let request = try makeDummyRequest()
+        let tokenStorage = OAuth2TokenMemoryStore()
+        var sut = OAuth2RequestPipelineMiddleware(clientConfiguration: validClientConfiguration, authorization: authorization, tokenStorage: tokenStorage)
+
+        sut.refreshStrategyFactory = nil
+
+        let tokenInvalidatedExpectation = expectation(description: "token refresh ignored & token invalidated")
+
+        let clientCredentialsStrategy = OAuth2ClientCredentialsTokenGrantStrategy(clientConfiguration: validClientConfiguration)
+        clientCredentialsStrategy.issueToken { result in
+            guard let token = result.value else {
+                XCTFail("No token")
+                return
+            }
+            let expiredToken = BearerToken(accessToken: token.accessToken, refreshToken: token.refreshToken, expiration: Date())
+
+            tokenStorage.store(token: expiredToken, for: validClientConfiguration, with: authorization)
+
+            sut.prepareForTransport(request: request) { result in
+                guard let error = result.error, case OAuth2Error.clientFailure(_, _) = error else {
+                    XCTFail("Custom strategy ignored")
+                    return
+                }
+
+                tokenInvalidatedExpectation.fulfill()
             }
         }
 
@@ -243,4 +334,50 @@ class OAuth2RequestPipelineMiddlewareTests: XCTestCase {
 
         waitForExpectations(timeout: 0.1)
     }
+
+    func testCoordinatesRefreshesBetweenMultipleSessions() throws {
+        /// Simulates multiple sessions (different processes) triggering token refreshes at once
+
+        let authorization = OAuth2Authorization(type: .bearer, level: .user)
+        let validClientConfiguration = try makeValidClientConfiguration()
+        let request = try makeDummyRequest()
+        let tokenStorage = OAuth2TokenMemoryStore()
+        let sut = OAuth2RequestPipelineMiddleware(clientConfiguration: validClientConfiguration, authorization: authorization, tokenStorage: tokenStorage)
+
+        let refreshTokenExpectation = expectation(description: "token refreshed")
+        let numSessions = 15
+        refreshTokenExpectation.expectedFulfillmentCount = numSessions
+        var authorizationHeaders: [String] = []
+        let arrayQueue = DispatchQueue(label: #function)
+
+        let clientCredentialsStrategy = OAuth2ClientCredentialsTokenGrantStrategy(clientConfiguration: validClientConfiguration)
+        let issuedToken = try clientCredentialsStrategy.issueToken()
+
+        let expiredToken = BearerToken(accessToken: issuedToken.accessToken, refreshToken: issuedToken.refreshToken, expiration: Date())
+
+        tokenStorage.store(token: expiredToken, for: validClientConfiguration, with: authorization)
+
+        for _ in 0..<numSessions {
+            sut.prepareForTransport(request: request) { result in
+                guard let request = result.value,
+                    let authorizationHeader = request.allHTTPHeaderFields?["Authorization"] else {
+                        XCTFail("No value")
+                        return
+                }
+                arrayQueue.sync {
+                    authorizationHeaders.append(authorizationHeader)
+                    refreshTokenExpectation.fulfill()
+                }
+            }
+        }
+
+        waitForExpectations(timeout: 5)
+
+        let firstHeader = authorizationHeaders[0]
+        XCTAssertFalse(firstHeader.contains(expiredToken.accessToken))
+        for header in authorizationHeaders {
+            XCTAssertEqual(firstHeader, header)
+        }
+    }
 }
+// swiftlint:enable type_body_length