+++ Tests/ConduitTests/Auth/OAuth2TokenStorageTests.swift
@@ -37,52 +37,123 @@ class OAuth2TokenStorageTests: XCTestCase {
         XCTAssertNil(storedToken)
     }
 
+    private func verifyRefreshTokenLockOperations<Token: OAuth2Token & DataConvertible>(sut: OAuth2TokenStore, with token: Token) throws {
+        let mockClientConfiguration = try makeMockClientConfiguration()
+        sut.unlockRefreshTokenFor(client: mockClientConfiguration, authorization: mockAuthorization)
+        XCTAssertFalse(sut.isRefreshTokenLockedFor(client: mockClientConfiguration, authorization: mockAuthorization))
+        sut.lockRefreshToken(timeout: 0.3, client: mockClientConfiguration, authorization: mockAuthorization)
+        XCTAssert(sut.isRefreshTokenLockedFor(client: mockClientConfiguration, authorization: mockAuthorization))
+
+        let waitExpectation = expectation(description: "refresh token unlocked")
+        DispatchQueue.global().asyncAfter(deadline: .now() + 0.31) {
+            XCTAssertFalse(sut.isRefreshTokenLockedFor(client: mockClientConfiguration, authorization: self.mockAuthorization))
+            waitExpectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 1)
+    }
+
     func testKeychainStorageOperations() throws {
         let sut = OAuth2TokenKeychainStore(service: UUID().uuidString)
         try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
     }
 
-    func testUserDefaultsStorageOperations() throws {
-        let sut = OAuth2TokenDiskStore(storageMethod: .userDefaults)
+    func testStandardUserDefaultsStorageOperations() throws {
+        let sut = OAuth2TokenUserDefaultsStore(userDefaults: .standard)
         try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
+    }
+
+    func testDomainUserDefaultsStorageOperations() throws {
+        let groupIdentifier = "group.com.mindbodyonline.ConduitTests"
+        guard let userDefaults = UserDefaults(suiteName: groupIdentifier) else {
+            XCTFail("Failed to create sandboxed application group (this won't work with codesigning)")
+            return
+        }
+        let sut = OAuth2TokenUserDefaultsStore(userDefaults: userDefaults)
+        try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
     }
 
 #if !os(tvOS)
     func testFileStorageOperations() throws {
-        let storagePath = NSTemporaryDirectory().appending(UUID().uuidString + ".oauth-token.bin")
+        let storagePath = NSTemporaryDirectory()
         let storageURL = URL(fileURLWithPath: storagePath)
-        let sut = OAuth2TokenDiskStore(storageMethod: .url(storageURL))
+        var sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL))
         try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
+        #if !os(macOS)
+        /// File protection options are unavailable on Mac
+        sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL,
+                                                                        coordinatesFileAccess: true,
+                                                                        tokenWritingOptions: [.atomic, .completeFileProtection]))
+        try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
+        #endif
+        sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL, coordinatesFileAccess: true))
+        try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
     }
 #endif
+
+    func testLegacyDiskStorageOperations() throws {
+        var sut = OAuth2TokenDiskStore(storageMethod: .userDefaults)
+        try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
+
+        let storagePath = NSTemporaryDirectory().appending(UUID().uuidString + ".oauth-token.bin")
+        let storageURL = URL(fileURLWithPath: storagePath)
+        #if !os(tvOS)
+        sut = OAuth2TokenDiskStore(storageMethod: .url(storageURL))
+        try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        /// Legacy file storage does not support refresh token locks
+        #endif
+    }
 
     func testMemoryStorageOperations() throws {
         let sut = OAuth2TokenMemoryStore()
         try verifyTokenStorageOperations(sut: sut, with: mockToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockToken)
     }
 
     func testLegacyKeychainStorageOperations() throws {
         let sut = OAuth2TokenKeychainStore(service: UUID().uuidString)
         try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
     }
 
     func testLegacyUserDefaultsStorageOperations() throws {
-        let sut = OAuth2TokenDiskStore(storageMethod: .userDefaults)
+        let sut = OAuth2TokenUserDefaultsStore(userDefaults: .standard)
         try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
     }
 
     #if !os(tvOS)
     func testLegacyFileStorageOperations() throws {
-        let storagePath = NSTemporaryDirectory().appending(UUID().uuidString + ".oauth-token.bin")
+        let storagePath = NSTemporaryDirectory()
         let storageURL = URL(fileURLWithPath: storagePath)
-        let sut = OAuth2TokenDiskStore(storageMethod: .url(storageURL))
+        var sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL))
         try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
+        #if !os(macOS)
+        /// File protection options are unavailable on Mac
+        sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL,
+                                                                        coordinatesFileAccess: true,
+                                                                        tokenWritingOptions: [.atomic, .completeFileProtection]))
+        try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
+        #endif
+        sut = OAuth2TokenFileStore(options: OAuth2TokenFileStoreOptions(storageDirectory: storageURL, coordinatesFileAccess: true))
+        try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
     }
     #endif
 
     func testLegacyMemoryStorageOperations() throws {
         let sut = OAuth2TokenMemoryStore()
         try verifyTokenStorageOperations(sut: sut, with: mockLegacyToken)
+        try verifyRefreshTokenLockOperations(sut: sut, with: mockLegacyToken)
     }
 
     private func validateLegacyTokenMigration(sut: OAuth2TokenStore) throws {